6th-term/Industrial-Programming/task3/todolist2/requirements.txt
Django==1.11.6
psycopg2-binary
djangorestframework
markdown
django-filter
django-cors-headers